chapter 4 ported

+++ src/include/vec3.cuh
@@ -9,20 +9,20 @@ using std::sqrt;
 
 class Vec3 {
 public:
-  double e[3];
+  float e[3];
 
   __host__ __device__ Vec3() : e{0, 0, 0} {}
-  __host__ __device__ Vec3(double e0, double e1, double e2) : e{e0, e1, e2} {}
+  __host__ __device__ Vec3(float e0, float e1, float e2) : e{e0, e1, e2} {}
 
-  __host__ __device__ double x() const { return e[0]; }
-  __host__ __device__ double y() const { return e[1]; }
-  __host__ __device__ double z() const { return e[2]; }
+  __host__ __device__ float x() const { return e[0]; }
+  __host__ __device__ float y() const { return e[1]; }
+  __host__ __device__ float z() const { return e[2]; }
 
   __host__ __device__ Vec3 operator-() const {
     return Vec3(-e[0], -e[1], -e[2]);
   }
-  __host__ __device__ double operator[](int i) const { return e[i]; }
-  __host__ __device__ double &operator[](int i) { return e[i]; }
+  __host__ __device__ float operator[](int i) const { return e[i]; }
+  __host__ __device__ float &operator[](int i) { return e[i]; }
 
   __host__ __device__ Vec3 &operator+=(const Vec3 &v) {
     e[0] += v.e[0];
@@ -31,20 +31,20 @@ public:
     return *this;
   }
 
-  __host__ __device__ Vec3 &operator*=(double t) {
+  __host__ __device__ Vec3 &operator*=(float t) {
     e[0] *= t;
     e[1] *= t;
     e[2] *= t;
     return *this;
   }
 
-  __host__ __device__ Vec3 &operator/=(double t) { return *this *= 1 / t; }
+  __host__ __device__ Vec3 &operator/=(float t) { return *this *= 1 / t; }
 
-  __host__ __device__ double lengthSquared() const {
+  __host__ __device__ float lengthSquared() const {
     return e[0] * e[0] + e[1] * e[1] + e[2] * e[2];
   }
 
-  __host__ __device__ double length() const { return sqrt(lengthSquared()); }
+  __host__ __device__ float length() const { return sqrt(lengthSquared()); }
 };
 
 // probably not the best for type-safety, but we will use it for geometric
